./2-strlen

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,9 +7,9 @@
 *Return: The length of @str
 */
 
-size_t _strlen(char *s)
+int _strlen(char *s)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*s++)
 	length++;
